+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Change Log
 
 v2.19.0
 ---
+* Fixed very rare cases when `rotateStringArray` couldn't rotate array properly
 * Improved `selfDefending` option
 * Fixed https://github.com/javascript-obfuscator/javascript-obfuscator/issues/959